admin/backend/pyproject.toml
[project]
name = "tnved-admin"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "surrealdb>=1.0",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py312"
